+++ jscomp/common/ast_config.mli
@@ -24,5 +24,5 @@
 
 val add_structure : string -> (Parsetree.expression option -> unit) -> unit
 val add_signature : string -> (Parsetree.expression option -> unit) -> unit
-val iter_on_bs_config_stru : Parsetree.structure -> unit
-val iter_on_bs_config_sigi : Parsetree.signature -> unit
+val iter_on_mel_config_stru : Parsetree.structure -> unit
+val iter_on_mel_config_sigi : Parsetree.signature -> unit