+++ jscomp/core/js_implementation.ml
@@ -15,13 +15,13 @@
 module Ppx_entry = struct
   let rewrite_signature (ast : Parsetree.signature) : Parsetree.signature =
     Bs_ast_invariant.iter_warnings_on_sigi ast;
-    Ast_config.iter_on_bs_config_sigi ast;
+    Ast_config.iter_on_mel_config_sigi ast;
     Bs_ast_invariant.emit_external_warnings_on_signature ast;
     ast
 
   let rewrite_implementation (ast : Parsetree.structure) : Parsetree.structure =
     Bs_ast_invariant.iter_warnings_on_stru ast;
-    Ast_config.iter_on_bs_config_stru ast;
+    Ast_config.iter_on_mel_config_stru ast;
     Bs_ast_invariant.emit_external_warnings_on_structure ast;
     ast
 end
@@ -77,7 +77,7 @@ let process_with_gentype filename =
       ignore (Sys.command comm)
 
 let after_parsing_sig ppf outputprefix ast =
-  Ast_config.iter_on_bs_config_sigi ast;
+  Ast_config.iter_on_mel_config_sigi ast;
   if !Js_config.modules then
     output_deps_set !Location.input_name
       (Ast_extract.read_parse_and_extract Mli ast);
@@ -161,7 +161,7 @@ let after_parsing_impl ppf fname (ast : Parsetree.structure) =
   let sourceintf = Filename.remove_extension fname ^ !Config.interface_suffix in
   Js_config.all_module_aliases :=
     (not (Sys.file_exists sourceintf)) && all_module_alias ast;
-  Ast_config.iter_on_bs_config_stru ast;
+  Ast_config.iter_on_mel_config_stru ast;
   let ast = if !Js_config.no_export then no_export ast else ast in
   if !Js_config.modules then
     output_deps_set !Location.input_name